pyproject.toml
[tool.poetry]
name = "mcodingbot"
version = "0.1.0"
description = "The Discord bot for the mCoding Discord server."
authors = ["Your Name <you@example.com>"]
license = "MIT"

[tool.poetry.scripts]
bot = "mcodingbot.__main__:main"

[tool.poetry.dependencies]
python = "^3.10,<3.11"
hikari = "^2.0.0.dev113"
hikari-crescent = "^0.4.0"
apgorm = "^1.0.0b12"
rapidfuzz = "^2.10.0"
cachetools = "^5.2.0"
types-cachetools = "^5.2.1"
floodgate-rs = "^0.1.1"
hikari-flare = "^0.1.0b3"

[tool.poetry.dev-dependencies]
nox = ">=2022.8.7"
black = "^22.8.0"
ruff = "^0.0.151"
isort = "^5.10.1"
mypy = "^0.991"

[tool.black]
line-length = 79
skip-magic-trailing-comma = true

[tool.ruff]
line-length = 79

[tool.isort]
profile = "black"

[tool.mypy]
warn_return_any = true
warn_unused_configs = true
strict = true
namespace_packages = true

[[tool.mypy.overrides]]
module = 'asyncpg.*'
ignore_missing_imports = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
